Refactored name to SendNotificationService
